# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OrangeTrack - Sistema de Gestão de Laranjas
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ORDERS
...
  /production/lots:
    get:
      operationId: listProductionLots
      tags: [production]
      summary: List production lots
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_production, finished, quality_approved, quality_rejected]
      responses:
        "200":
          description: Production lots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionLot"
    post:
      operationId: createProductionLot
      tags: [production]
      summary: Create a production lot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductionLotBody"
      responses:
        "201":
          description: Created lot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionLot"
  /production/lots/{id}:
    get:
      operationId: getProductionLot
      tags: [production]
      summary: Get production lot by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Production lot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionLot"
    put:
      operationId: updateProductionLot
      tags: [production]
      summary: Update production lot status or quality data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductionLotBody"
      responses:
        "200":
          description: Updated lot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionLot"
...
  /reports/production:
    get:
      operationId: getProductionReport
      tags: [reports]
      summary: Get production report
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Production report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionReport"
  /reports/financial:
    get:
      operationId: getFinancialReport
      tags: [reports]
      summary: Get financial report
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Financial report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialReport"
  /reports/quality:
    get:
      operationId: getQualityReport
      tags: [reports]
      summary: Get quality control report
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Quality report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityReport"
  /reports/discards:
    get:
      operationId: getDiscardReport
      tags: [reports]
      summary: Get discard report
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Discard report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscardReport"
components:
  schemas:
...
    ProductionChartPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        juiceKg:
          type: number
        orangeKg:
          type: number
      required: [date, juiceKg, orangeKg]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [juice, fruit, byproduct]
        unit:
          type: string
          enum: [kg, liter, unit, box]
        pricePerUnit:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, unit, pricePerUnit, createdAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [juice, fruit, byproduct]
        unit:
          type: string
          enum: [kg, liter, unit, box]
        pricePerUnit:
          type: number
        description:
          type: string
      required: [name, type, unit, pricePerUnit]

    # Order
...
    ProductionLot:
      type: object
      properties:
        id:
          type: integer
        lotCode:
          type: string
        productId:
          type: integer
        productName:
          type: string
        orderId:
          type: integer
        rawMaterialId:
          type: integer
        quantityProduced:
          type: number
        unit:
          type: string
        status:
          type: string
          enum: [pending, in_production, finished, quality_approved, quality_rejected]
        qualityStatus:
          type: string
          enum: [pending, approved, rejected]
        brixLevel:
          type: number
        temperature:
          type: number
        qualityNotes:
          type: string
        producedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, lotCode, productId, productName, quantityProduced, unit, status, qualityStatus, createdAt]
    CreateProductionLotBody:
      type: object
      properties:
        productId:
          type: integer
        orderId:
          type: integer
        rawMaterialId:
          type: integer
        quantityProduced:
          type: number
        unit:
          type: string
      required: [productId, quantityProduced, unit]
    UpdateProductionLotBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_production, finished, quality_approved, quality_rejected]
        qualityStatus:
          type: string
          enum: [pending, approved, rejected]
        brixLevel:
          type: number
        temperature:
          type: number
        qualityNotes:
          type: string
...
    ProductionReport:
      type: object
      properties:
        totalProduced:
          type: number
        totalLots:
          type: integer
        approvedLots:
          type: integer
        rejectedLots:
          type: integer
        averageBrix:
          type: number
        byProduct:
          type: array
          items:
            type: object
            properties:
              productName:
                type: string
              quantity:
                type: number
              unit:
                type: string
            required: [productName, quantity, unit]
        dailyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ProductionChartPoint"
      required: [totalProduced, totalLots, approvedLots, rejectedLots, averageBrix, byProduct, dailyBreakdown]
    FinancialReport:
      type: object
      properties:
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        averageOrderValue:
          type: number
        byPaymentType:
          type: array
          items:
            type: object
            properties:
              paymentType:
                type: string
              total:
                type: number
              count:
                type: integer
            required: [paymentType, total, count]
        topCustomers:
          type: array
          items:
            type: object
            properties:
              customerName:
                type: string
              totalSpent:
                type: number
            required: [customerName, totalSpent]
      required: [totalRevenue, totalOrders, averageOrderValue, byPaymentType, topCustomers]
    QualityReport:
      type: object
      properties:
        totalLotsTested:
          type: integer
        approvedLots:
          type: integer
        rejectedLots:
          type: integer
        approvalRate:
          type: number
        averageBrix:
          type: number
        averageTemperature:
          type: number
        rejectionReasons:
          type: array
          items:
            type: object
            properties:
              reason:
                type: string
              count:
                type: integer
            required: [reason, count]
      required:
        - totalLotsTested
        - approvedLots
        - rejectedLots
        - approvalRate
        - averageBrix
        - averageTemperature
        - rejectionReasons
    DiscardReport:
      type: object
      properties:
        totalDiscardedKg:
          type: number
        totalMovements:
          type: integer
        byReason:
          type: array
          items:
            type: object
            properties:
              reason:
                type: string
              quantityKg:
                type: number
              count:
                type: integer
            required: [reason, quantityKg, count]
      required: [totalDiscardedKg, totalMovements, byReason]