# lib/api-zod/src/generated/types/productionLot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OrangeTrack - Sistema de Gestão de Laranjas
 * OpenAPI spec version: 0.1.0
 */
import type { ProductionLotQualityStatus } from "./productionLotQualityStatus";
import type { ProductionLotStatus } from "./productionLotStatus";

export interface ProductionLot {
  id: number;
  lotCode: string;
  productId: number;
  productName: string;
  orderId?: number;
  rawMaterialId?: number;
  quantityProduced: number;
  unit: string;
  status: ProductionLotStatus;
  qualityStatus: ProductionLotQualityStatus;
  brixLevel?: number;
  temperature?: number;
  qualityNotes?: string;
  producedAt?: Date;
  createdAt: Date;
}